NotificationServiceApi/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
dnspython==2.4.2
drf-yasg==1.21.7
eventlet==0.33.3
flower==2.0.1
greenlet==2.0.2
humanize==4.8.0
idna==3.4
inflection==0.5.1
kombu==5.3.2
packaging==23.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tornado==6.3.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
